+++ Scroller/EditableWrapper.js
@@ -87,7 +87,6 @@ const EditableWrapper = (props) => {
 	const hideItemFuncRef = editable?.hideItemFuncRef;
 	const showItemFuncRef = editable?.showItemFuncRef;
 	const focusItemFuncRef = editable?.focusItemFuncRef;
-	const initialSelected = editable?.initialSelected;
 	const mergedCss = usePublicClassNames({componentCss, customCss, publicClassNames: true});
 
 	const dataSize = children?.length;
@@ -131,7 +130,10 @@ const EditableWrapper = (props) => {
 		// Flag for prevent event propagation
 		needToPreventEvent: null,
 
-		lastInputDirection: null
+		lastInputDirection: null,
+
+		// initialSelected
+		initialSelected: editable?.initialSelected
 	});
 	const announceRef = useRef({});
 
@@ -151,11 +153,11 @@ const EditableWrapper = (props) => {
 		mutableRef.current.selectedItemLabel = '';
 		mutableRef.current.lastMoveDirection = null;
 		mutableRef.current.prevToIndex = null;
-		initialSelected.itemIndex = null;
+		mutableRef.current.initialSelected = null;
 		wrapperRef.current.style.setProperty('--selected-item-offset', '0px');
 
 		Spotlight.set(spotlightId, {restrict: 'self-first'});
-	}, [customCss.focused, customCss.selected, initialSelected]);
+	}, [customCss.focused, customCss.selected]);
 
 	// Finalize the order
 	const finalizeOrders = useCallback(() => {
@@ -213,7 +215,7 @@ const EditableWrapper = (props) => {
 			mutableRef.current.prevToIndex = mutableRef.current.fromIndex;
 
 			updateArrowIcon(mutableRef.current.fromIndex);
-			if (!initialSelected?.itemIndex) {
+			if (!mutableRef.current.initialSelected) {
 				setTimeout(() => {
 					announceRef.current.announce(
 						mutableRef.current.selectedItemLabel + $L('Press the left/right button to move or press the up button to select other options.')
@@ -221,15 +223,15 @@ const EditableWrapper = (props) => {
 				}, completeAnnounceDelay);
 			}
 		}
-	}, [customCss.focused, customCss.selected, initialSelected?.itemIndex, updateArrowIcon]);
+	}, [customCss.focused, customCss.selected, updateArrowIcon]);
 
 	const finalizeEditing = useCallback((orders) => {
-		if (initialSelected?.itemIndex) {
+		if (mutableRef.current.initialSelected) {
 			mutableRef.current.selectedItem.children[1].ariaLabel = `${mutableRef.current.selectedItem.ariaLabel} ${$L('Press the OK button to move or press the up button to select other options.')}`;
 		}
 		forwardCustom('onComplete', () => ({orders, hideIndex: mutableRef.current.hideIndex}))(null, editable);
 		reset();
-	}, [editable, initialSelected?.itemIndex, reset]);
+	}, [editable, reset]);
 
 	const findItemNode = useCallback((node) => {
 		for (let current = node; current !== scrollContentRef.current && current !== document; current = current.parentNode) {
@@ -754,15 +756,21 @@ const EditableWrapper = (props) => {
 	}, [getNextIndexFromPosition, moveItems, scrollContainerHandle, scrollContentRef]);
 
 	useEffect(() => {
-		if (initialSelected?.itemIndex) {
-			scrollContainerHandle.current?.scrollTo({animate:false, position: {x: initialSelected.scrollLeft}});
+		if (mutableRef.current.initialSelected) {
+			scrollContainerHandle.current?.scrollTo({animate:false, position: {x: mutableRef.current.initialSelected.scrollLeft}});
 		}
-	}, [initialSelected?.itemIndex, initialSelected?.scrollLeft, scrollContainerHandle]);
+	}, [scrollContainerHandle]);
 
 	useLayoutEffect(() => {
 		const iconItemList = Array.from(wrapperRef.current.children);
+		let initialSelected = mutableRef.current.initialSelected;
+
+		if (initialSelected && !(initialSelected?.itemIndex > 0)) { // filter nullish values
+			initialSelected = mutableRef.current.initialSelected = null;
+		}
+
 		if (initialSelected?.itemIndex) {
-			const initialSelectedItem = wrapperRef.current.children[initialSelected.itemIndex - 1];
+			const initialSelectedItem = wrapperRef.current.children[initialSelected?.itemIndex - 1];
 			if (initialSelectedItem?.dataset.index) {
 				mutableRef.current.focusedItem = initialSelectedItem;
 				mutableRef.current.lastMouseClientX = getLastPointerPosition().x;
@@ -774,7 +782,11 @@ const EditableWrapper = (props) => {
 
 		iconItemList.forEach((iconItemWrapper, index) => {
 			if (iconItemWrapper?.children[1]) {
-				iconItemWrapper.children[1].ariaLabel += index === initialSelected?.itemIndex - 1 ? ` ${$L('Press the left/right button to move or press the up button to select other options.')}` : ` ${$L('Press the OK button to move or press the up button to select other options.')}`;
+				if (initialSelected && (initialSelected.itemIndex - 1 === index)) {
+					iconItemWrapper.children[1].ariaLabel += ` ${$L('Press the left/right button to move or press the up button to select other options.')}`;
+				} else {
+					iconItemWrapper.children[1].ariaLabel += ` ${$L('Press the OK button to move or press the up button to select other options.')}`;
+				}
 			}
 		});
 	}, []); // eslint-disable-line react-hooks/exhaustive-deps